feat: Api v3 fileprovider

+++ kDriveFileProvider/FileProviderExtension+Actions.swift
@@ -35,11 +35,10 @@ extension FileProviderExtension {
 
             // Call completion handler with error if the file name already exists
             let itemsWithSameParent = file.children
-                .map { FileProviderItem(file: $0, domain: self.domain) } + self.fileProviderState
-                .importedDocuments(forParent: parentItemIdentifier)
+                .map { $0.toFileProviderItem(parent: nil, domain: self.domain) }
             let newItemFileName = directoryName.lowercased()
             if let collidingItem = itemsWithSameParent.first(where: { $0.filename.lowercased() == newItemFileName }),
-               !collidingItem.isTrashed {
+               !(collidingItem.isTrashed ?? false) {
                 completionHandler(nil, NSError.fileProviderErrorForCollision(with: collidingItem))
                 return
             }
@@ -51,7 +50,7 @@ extension FileProviderExtension {
                     name: directoryName,
                     onlyForMe: false
                 )
-                completionHandler(FileProviderItem(file: directory, domain: self.domain), nil)
+                completionHandler(directory.toFileProviderItem(parent: nil, domain: self.domain), nil)
             } catch {
                 completionHandler(nil, error)
             }
@@ -65,7 +64,7 @@ extension FileProviderExtension {
         Log.fileProvider("deleteItem")
         Task {
             guard let fileId = itemIdentifier.toFileId() else {
-                completionHandler(self.nsError(code: .noSuchItem))
+                completionHandler(NSFileProviderError(.noSuchItem))
                 return
             }
 
@@ -74,11 +73,13 @@ extension FileProviderExtension {
                     .deleteDefinitely(file: ProxyFile(driveId: self.driveFileManager.drive.id, id: fileId))
                 if response {
                     self.fileProviderState.removeWorkingDocument(forKey: itemIdentifier)
+                    completionHandler(nil)
+
+                    // Signal after completionHandler
                     try await self.manager.signalEnumerator(for: .workingSet)
                     try await self.manager.signalEnumerator(for: itemIdentifier)
-                    completionHandler(nil)
                 } else {
-                    completionHandler(self.nsError(code: .serverUnreachable))
+                    completionHandler(NSFileProviderError(.serverUnreachable))
                 }
             } catch {
                 completionHandler(error)
@@ -101,18 +102,17 @@ extension FileProviderExtension {
             return
         }
         let itemsWithSameParent = file.children
-            .map { FileProviderItem(file: $0, domain: self.domain) } + fileProviderState
-            .importedDocuments(forParent: parentItemIdentifier)
+            .map { $0.toFileProviderItem(parent: nil, domain: self.domain) }
         let newItemFileName = fileURL.lastPathComponent.lowercased()
         if let collidingItem = itemsWithSameParent.first(where: { $0.filename.lowercased() == newItemFileName }),
-           !collidingItem.isTrashed {
+           !(collidingItem.isTrashed ?? false) {
             completionHandler(nil, NSError.fileProviderErrorForCollision(with: collidingItem))
             return
         }
 
-        let id = UUID().uuidString
-        let importedDocumentIdentifier = NSFileProviderItemIdentifier(id)
-        let storageUrl = FileProviderItem.createStorageUrl(
+        let importedFileUUID = UUID().uuidString
+        let importedDocumentIdentifier = NSFileProviderItemIdentifier(importedFileUUID)
+        let storageUrl = fileProviderService.createStorageUrl(
             identifier: importedDocumentIdentifier,
             filename: fileURL.lastPathComponent,
             domain: domain
@@ -134,17 +134,20 @@ extension FileProviderExtension {
         if accessingSecurityScopedResource {
             fileURL.stopAccessingSecurityScopedResource()
         }
-        let importedItem = FileProviderItem(
-            importedFileUrl: storageUrl,
-            identifier: importedDocumentIdentifier,
-            parentIdentifier: parentItemIdentifier
-        )
-        fileProviderState.setImportedDocument(importedItem, forKey: importedDocumentIdentifier)
 
-        backgroundUploadItem(importedItem)
+        guard let parentDirectoryId = parentItemIdentifier.toFileId() else {
+            fatalError("missing parentDirectoryId TODO sentry")
+        }
 
-        manager.signalEnumerator(for: parentItemIdentifier) { _ in
-            completionHandler(importedItem, nil)
+        let importItem = UploadFileProviderItem(uploadFileUUID: importedFileUUID,
+                                                parentDirectoryId: parentDirectoryId,
+                                                userId: driveFileManager.drive.userId,
+                                                driveId: driveFileManager.drive.id,
+                                                sourceUrl: storageUrl,
+                                                conflictOption: .version,
+                                                driveError: nil)
+        backgroundUpload(importItem) {
+            completionHandler(importItem, nil)
         }
     }
 
@@ -155,11 +158,9 @@ extension FileProviderExtension {
     ) {
         Log.fileProvider("renameItem")
         Task {
-            // Doc says we should do network request after renaming local file but we could end up with model desync
-            if let item = self.fileProviderState.getImportedDocument(forKey: itemIdentifier) {
-                item.filename = itemName
-                try await self.manager.signalEnumerator(for: item.parentItemIdentifier)
-                completionHandler(item, nil)
+            guard self.uploadQueue.getUploadingFile(fileProviderItemIdentifier: itemIdentifier.rawValue) == nil else {
+                Log.fileProvider("renameItem not supported while uploading", level: .error)
+                completionHandler(nil, NSFileProviderError(.noSuchItem))
                 return
             }
 
@@ -170,13 +171,11 @@ extension FileProviderExtension {
             }
 
             // Check if file name already exists
-            let item = FileProviderItem(file: file, domain: self.domain)
             let itemsWithSameParent = file.parent!.children
-                .map { FileProviderItem(file: $0, domain: self.domain) } + self.fileProviderState
-                .importedDocuments(forParent: item.parentItemIdentifier)
+                .map { $0.toFileProviderItem(parent: nil, domain: self.domain) }
             let newItemFileName = itemName.lowercased()
             if let collidingItem = itemsWithSameParent.first(where: { $0.filename.lowercased() == newItemFileName }),
-               !collidingItem.isTrashed {
+               !(collidingItem.isTrashed ?? false) {
                 completionHandler(nil, NSError.fileProviderErrorForCollision(with: collidingItem))
                 return
             }
@@ -184,7 +183,9 @@ extension FileProviderExtension {
             let proxyFile = file.proxify()
             do {
                 let file = try await self.driveFileManager.rename(file: proxyFile, newName: itemName)
-                completionHandler(FileProviderItem(file: file.freeze(), domain: self.domain), nil)
+                completionHandler(
+                    file.freeze().toFileProviderItem(parent: nil, domain: self.domain), nil
+                )
             } catch {
                 completionHandler(nil, error)
             }
@@ -199,10 +200,9 @@ extension FileProviderExtension {
     ) {
         Log.fileProvider("reparentItem")
         Task {
-            if let item = self.fileProviderState.getImportedDocument(forKey: itemIdentifier) {
-                item.parentItemIdentifier = parentItemIdentifier
-                try await self.manager.signalEnumerator(for: item.parentItemIdentifier)
-                completionHandler(item, nil)
+            guard self.uploadQueue.getUploadingFile(fileProviderItemIdentifier: itemIdentifier.rawValue) == nil else {
+                Log.fileProvider("reparentItem not supported while uploading", level: .error)
+                completionHandler(nil, NSFileProviderError(.noSuchItem))
                 return
             }
 
@@ -218,7 +218,9 @@ extension FileProviderExtension {
             let proxyParent = parent.proxify()
             do {
                 let (_, file) = try await self.driveFileManager.move(file: proxyFile, to: proxyParent)
-                completionHandler(FileProviderItem(file: file.freeze(), domain: self.domain), nil)
+                completionHandler(
+                    file.freeze().toFileProviderItem(parent: nil, domain: self.domain), nil
+                )
             } catch {
                 completionHandler(nil, error)
             }
@@ -231,7 +233,7 @@ extension FileProviderExtension {
         completionHandler: @escaping (NSFileProviderItem?, Error?) -> Void
     ) {
         let fileId = itemIdentifier.toFileId()
-        Log.fileProvider("setFavoriteRank forItemIdentifier:\(fileId)")
+        Log.fileProvider("setFavoriteRank forItemIdentifier:\(String(describing: fileId))")
         // How should we save favourite rank in database?
         guard let fileId,
               let file = driveFileManager.getCachedFile(id: fileId) else {
@@ -239,8 +241,8 @@ extension FileProviderExtension {
             return
         }
 
-        let item = FileProviderItem(file: file, domain: domain)
-        item.favoriteRank = favoriteRank
+        let item = file.toFileProviderItem(parent: nil, domain: domain)
+        item.favoriteRankModifier(newValue: favoriteRank)
 
         completionHandler(item, nil)
     }
@@ -252,7 +254,7 @@ extension FileProviderExtension {
     ) {
         Log.fileProvider("setLastUsedDate forItemIdentifier")
         // kDrive doesn't support this
-        completionHandler(nil, NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError))
+        completionHandler(nil, NSError.featureUnsupported)
     }
 
     override func setTagData(
@@ -262,7 +264,7 @@ extension FileProviderExtension {
     ) {
         Log.fileProvider("setTagData :\(tagData?.count) forItemIdentifier")
         // kDrive doesn't support this
-        completionHandler(nil, NSError(domain: NSCocoaErrorDomain, code: NSFeatureUnsupportedError))
+        completionHandler(nil, NSError.featureUnsupported)
     }
 
     override func trashItem(
@@ -284,11 +286,16 @@ extension FileProviderExtension {
 
             // Make deleted file copy
             let deletedFile = file.detached()
-            let item = FileProviderItem(file: deletedFile, domain: self.domain)
-            item.isTrashed = true
+            let item = deletedFile.toFileProviderItem(parent: nil, domain: self.domain)
+            item.trashModifier(newValue: true)
+
             let proxyFile = file.proxify()
 
             do {
+                guard let item = item as? FileProviderItem else {
+                    return
+                }
+
                 _ = try await self.driveFileManager.delete(file: proxyFile)
                 self.fileProviderState.setWorkingDocument(item, forKey: itemIdentifier)
                 completionHandler(item, nil)
@@ -307,7 +314,7 @@ extension FileProviderExtension {
         Log.fileProvider("untrashItem withIdentifier:\(fileId)")
         Task {
             guard let fileId else {
-                completionHandler(nil, self.nsError(code: .noSuchItem))
+                completionHandler(nil, NSFileProviderError(.noSuchItem))
                 return
             }
 
@@ -323,17 +330,17 @@ extension FileProviderExtension {
                 }
                 // Restore in given parent
                 _ = try await self.driveFileManager.apiFetcher.restore(file: file.proxify(), in: parent)
-                let item = FileProviderItem(file: file, domain: self.domain)
-                if let parentItemIdentifier {
-                    item.parentItemIdentifier = parentItemIdentifier
-                }
-                item.isTrashed = false
+                let item = file.toFileProviderItem(parent: parentItemIdentifier, domain: self.domain)
+                item.trashModifier(newValue: false)
+
                 self.fileProviderState.removeWorkingDocument(forKey: itemIdentifier)
+                completionHandler(item, nil)
+
+                // Signal after completionHandler
                 try await self.manager.signalEnumerator(for: .workingSet)
                 try await self.manager.signalEnumerator(for: item.parentItemIdentifier)
-                completionHandler(item, nil)
             } catch {
-                completionHandler(nil, self.nsError(code: .noSuchItem))
+                completionHandler(nil, NSFileProviderError(.noSuchItem))
             }
         }
     }
